Bootcamps/BairesDev - Machine Learning Training/Criação de Uma Base de Dados e Treinamento da Rede YOLO/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

Pillow==11.3.0
ultralytics
pycocotools
torch
torchvision